extract last error out of uniques
